+++ R/interactive_example.R
@@ -20,7 +20,7 @@ fetch_terms_for_example <- function(de) {
   # load Reactome pathways
   message("Fetching Reactome data\n")
   re <- fetch_reactome("Saccharomyces cerevisiae", on_error = "warn")
-  re_data <- prepare_for_enrichment(re$terms, re$mapping, all_genes, feature_name = "gene_id")
+  re_data <- prepare_for_enrichment(re$terms, re$mapping, all_genes, feature_name = "ensembl_id")
 
   # Put all functional term data in one structure; Shiny app will access
   # individual ontologies from this list